docs: Group top-level reflections
