--- examples/index.md
@@ -1,16 +1,0 @@
----
-title: Examples
-category: Examples
-children:
-- ./console.md
-- ./download-code.md
-- ./download-memory.md
-- ./socket-demo.md
-- ./terminal-client.md
----
-**Warning:** These examples may be outdated and plain broken.
-
-If you've cloned this repository and installed the dev dependencies, you can run these demo scripts from the command line using `tsx`:
-```sh
-yarn exec tsx examples/example-script-name.ts
-```
