--- guides/index.md
@@ -1,12 +1,0 @@
----
-title: Guides
-category: Guides
-children:
-- ./configuration.md
-- ./debugging.md
-- ./migration-2.md
-- ./rate-limits.md
-- ./servers.md
-- ./websocket.md
----
-These documents provide in-depth coverage of specific topics.
