+++ src/common/index.ts
@@ -3,9 +3,3 @@ export * from './market'
 export * from './resources'
 export * from './rooms'
 export * from './users'
-
-/**
- * Defines types/constants/enums/etc that are relevant to both
- * the HTTP and WebSocket APIs
- * @module
- */
