+++ src/common/users.ts
@@ -1,21 +1,23 @@
 /**
- * Defines types/constants/enums/etc related to users and their accounts
- * @module
+ * High-level metadata for an individual user
+ * @category Common - Users
  */
-
-/** High-level metadata for an individual user */
 export interface User {
   _id: string
   username: string
   badge: UserBadge
 }
 
-/** A collection of user metadata keyed by user ID */
+/**
+ * A collection of user metadata keyed by user ID
+ * @category Common - Users
+ */
 export interface Users { [userId: string]: User }
 
 /**
  * Parameters used to render a player's SVG icon / logo / "bust"
  * (as it is referenced in the client)
+ * @category Common - Users
  */
 export interface UserBadge {
   /** Primary badge color (in a web color format) */
@@ -41,13 +43,17 @@ export interface UserBadge {
 /**
  * A pair of SVG paths composing a premium {@link UserBadge}
  * @see {BadgeDecoration}
+ * @category Common - Users
  */
 export interface UserBadgeSvgs {
   path1: string
   path2: string
 }
 
-/** A collection of JavaScript/WASM modules */
+/**
+ * A collection of JavaScript/WASM modules
+ * @category Common - Users
+ */
 export interface UserCodeModules {
   /** JavaScript code or WASM binaries keyed by module name. */
   [moduleName: string]: UserCodeModule
@@ -59,10 +65,14 @@ export interface UserCodeModules {
  * If the value is a string, the module is JavaScript code.
  * If the value has a `binary` property, the module is a {@link UserCodeWasmModule}.
  * @see {@link UserCodeModules}
+ * @category Common - Users
  */
 export type UserCodeModule = string | UserCodeWasmModule
 
-/** Binary contents of a WebAssembly (WASM) module */
+/**
+ * Binary contents of a WebAssembly (WASM) module
+ * @category Common - Users
+ */
 export interface UserCodeWasmModule {
   binary: string
 }
@@ -71,11 +81,18 @@ export interface UserCodeWasmModule {
  * A user's CPU limits for each shard indexed by name.
  *
  * `undefined` is equivalent to 0.
+ * @category Common - Users
  */
 export interface CpuShardLimits { [shardName: string]: number | undefined }
 
-/** ID of the "{@link https://docs.screeps.com/invaders.html | Invader}" NPC */
+/**
+ * ID of the "{@link https://docs.screeps.com/invaders.html | Invader}" NPC
+ * @category Common - Users
+ */
 export const INVADER_ID = '2'
 
-/** ID of the "Source Keeper" NPC */
+/**
+ * ID of the "Source Keeper" NPC
+ * @category Common - Users
+ */
 export const SOURCE_KEEPER_ID = '3'
