+++ src/http/gameMarket.ts
@@ -3,13 +3,9 @@ import { MarketResource } from '../common/resources'
 import { ScreepsResponse } from './base'
 
 /**
- * Used with HTTP API endpoints on the `/api/game/market` path
- * @module
- */
-
-/**
  * `GET /api/game/market/orders-index` response
  * @see {@link ScreepsHttpClient.gameMarketOrdersIndex}
+ * @category HTTP API - Game/Market
  */
 export interface GameMarketIndexResponse extends ScreepsResponse {
   list: {
@@ -26,6 +22,7 @@ export interface GameMarketIndexResponse extends ScreepsResponse {
  *
  * Intershard orders will be listed under the `intershard` key
  * @see {@link ScreepsHttpClient.gameMarketMyOrders}
+ * @category HTTP API - Game/Market
  */
 export type GameMarketMyOrdersResponse = ScreepsResponse & {
   [shardName: string]: Order[]
@@ -34,6 +31,7 @@ export type GameMarketMyOrdersResponse = ScreepsResponse & {
 /**
  * `GET /api/game/market/orders` response
  * @see {@link ScreepsHttpClient.gameMarketOrders}
+ * @category HTTP API - Game/Market
  */
 export interface GameMarketOrdersResponse extends ScreepsResponse {
   list: OpenOrder[]
@@ -42,6 +40,7 @@ export interface GameMarketOrdersResponse extends ScreepsResponse {
 /**
  * `GET /api/game/market/stats` resonse
  * @see {@link ScreepsHttpClient.gameMarketStats}
+ * @category HTTP API - Game/Market
  */
 export interface GameMarketStatsResponse extends ScreepsResponse {
   stats: {
