+++ src/http/index.ts
@@ -1,12 +1,10 @@
 export * from './auth'
 export * from './base'
-export * from './userCode'
 export * from './userDecorations'
 export * from './experimental'
 export * from './game'
 export * from './leaderboard'
 export * from './gameMarket'
-export * from './userMemory'
 export * from './userMessages'
 export * from './scoreboard'
 export * from './seasons'
@@ -14,9 +12,3 @@ export * from './servers'
 export * from './gameShards'
 export * from './user'
 export * from './warpath'
-
-/**
- * Defines types/constants/enums/etc related to the HTTP API
- * @see {@link ScreepsHttpClient}
- * @module
- */
