+++ src/http/leaderboard.ts
@@ -2,13 +2,9 @@ import { Users } from '../common/users'
 import { ScreepsResponse } from './base'
 
 /**
- * Used with HTTP API endpoints on the `/api/leaderboard` path
- * @module
- */
-
-/**
  * `GET /api/leaderboard/list` response
  * @see {@link ScreepsHttpClient.leaderboardList}
+ * @category HTTP API - Leaderboard
  */
 export interface LeaderboardListResponse extends ScreepsResponse {
   list: LeaderboardResult[]
@@ -20,10 +16,14 @@ export interface LeaderboardListResponse extends ScreepsResponse {
 /**
  * `GET /api/leaderboard/find` response
  * @see {@link ScreepsHttpClient.leaderboardFind}
+ * @category HTTP API - Leaderboard
  */
 export interface LeaderboardFindResponse extends ScreepsResponse, LeaderboardResult {}
 
-/** A user's leaderboard result for a single season */
+/**
+ * A user's leaderboard result for a single season
+ * @category HTTP API - Leaderboard
+ */
 export interface LeaderboardResult {
   _id: string
   season: string
@@ -35,6 +35,7 @@ export interface LeaderboardResult {
 /**
  * `GET /api/leaderboard/seasons` response
  * @see {@link ScreepsHttpClient.leaderboardSeasons}
+ * @category HTTP API - Leaderboard
  */
 export interface LeaderboardSeasonsResponse extends ScreepsResponse {
   seasons: {
@@ -51,6 +52,7 @@ export interface LeaderboardSeasonsResponse extends ScreepsResponse {
  * The types of leaderboard available via the leaderboard endpoints
  * (ex: {@link ScreepsHttpClient.leaderboardList})
  * @enum
+ * @category HTTP API - Leaderboard
  */
 export const LeaderboardModes = {
   /** "Power Processed" (as used to earn Global Power Level) leaderboard */
@@ -59,5 +61,8 @@ export const LeaderboardModes = {
   World: 'world'
 } as const
 
-/** A {@link LeaderboardModes} value */
+/**
+ * A {@link LeaderboardModes} value
+ * @category HTTP API - Leaderboard
+ */
 export type LeaderboardMode = typeof LeaderboardModes[keyof typeof LeaderboardModes]
