+++ src/http/servers.ts
@@ -1,19 +1,18 @@
 import { ScreepsResponse } from './base'
 
 /**
- * Used with HTTP API endpoints on the `/api/servers` path
- * @module
- */
-
-/**
  * `POST /api/servers/list` response: a curated list of community-run servers
  * @see {@link ScreepsHttpClient.serversList}
+ * @category HTTP API - Servers
  */
 export interface ServerListResponse extends ScreepsResponse {
   servers: Server[]
 }
 
-/** A server from {@link ServerListResponse} */
+/**
+ * A server from {@link ServerListResponse}
+ * @category HTTP API - Servers
+ */
 export interface Server {
   _id: string
   settings: {
