+++ src/http/user.ts
@@ -1,16 +1,12 @@
 import { MarketResource } from '../common/resources'
 import { RoomStat } from '../common/rooms'
-import { User } from '../common/users'
-import { ScreepsResponse } from './base'
-
-/**
- * Used with HTTP API endpoints on the `/api/user` path
- * @module
- */
+import { User, UserCodeModules } from '../common/users'
+import { ScreepsDbUpdateResponse, ScreepsResponse } from './base'
 
 /**
  * `POST /api/user/notify-prefs` request
  * @see {@link ScreepsHttpClient.userNotifyPrefs}
+ * @category HTTP API - User
  */
 export interface UserNotifyPrefsRequest {
   disabled: boolean
@@ -23,6 +19,7 @@ export interface UserNotifyPrefsRequest {
 /**
  * `GET /api/user/world-start-room` response
  * @see {@link ScreepsHttpClient.userWorldStartRoom}
+ * @category HTTP API - User
  */
 export interface UserWorldStartRoomResponse extends ScreepsResponse {
   /**
@@ -35,6 +32,7 @@ export interface UserWorldStartRoomResponse extends ScreepsResponse {
 /**
  * `GET /api/user/world-status` response
  * @see {@link ScreepsHttpClient.userWorldStatus}
+ * @category HTTP API - User
  */
 export interface UserWorldStatusResponse extends ScreepsResponse {
   status: UserWorldStatus
@@ -44,6 +42,7 @@ export interface UserWorldStatusResponse extends ScreepsResponse {
  * A user's status on the server. This is used to determine whether
  * or not they may claim a room or place a spawn directly (to spawn/respawn).
  * @enum
+ * @category HTTP API - User
  */
 export const UserWorldStatuses = {
   /** User has just entered the world or respawned and has yet to place a spawn */
@@ -57,12 +56,16 @@ export const UserWorldStatuses = {
   Lost: 'lost'
 } as const
 
-/** Any {@link UserWorldStatuses} value */
+/**
+ * Any {@link UserWorldStatuses} value
+ * @category HTTP API - User
+ */
 export type UserWorldStatus = typeof UserWorldStatuses[keyof typeof UserWorldStatuses]
 
 /**
  * `GET /api/user/branches` response
  * @see {@link ScreepsHttpClient.userBranches}
+ * @category HTTP API - User
  */
 export interface UserBranchesResponse extends ScreepsResponse {
   list: {
@@ -74,15 +77,41 @@ export interface UserBranchesResponse extends ScreepsResponse {
 }
 
 /**
+ * `GET /api/user/code` response
+ * @see {@link ScreepsHttpClient.userCodeGet}
+ * @category HTTP API - User
+ */
+export interface UserCodeGetResponse extends ScreepsResponse, UserCodeSetRequest {}
+
+/**
+ * `POST /api/user/code` response
+ * @see {@link ScreepsHttpClient.userCodeSet}
+ * @category HTTP API - User
+ */
+export interface UserCodeSetRequest {
+  /**
+   * The name of the branch
+   * @see {@link ScreepsHttpClient.userBranches} to list available branches
+   */
+  branch: string
+  /** JavaScript code and WASM binaries keyed by module name */
+  modules: UserCodeModules
+}
+
+/**
  * `GET /api/user/find` response
  * @see {@link ScreepsHttpClient.userFind}
  * @see {@link ScreepsHttpClient.userFindById}
+ * @category HTTP API - User
  */
 export interface UserFindResponse extends ScreepsResponse {
   user: UserInfo
 }
 
-/** A result from {@link UserFindResponse} */
+/**
+ * A result from {@link UserFindResponse}
+ * @category HTTP API - User
+ */
 export interface UserInfo extends User {
   /**
    * Total Global Control Level (GCL) progress/points earned by this user.
@@ -106,6 +135,7 @@ export interface UserInfo extends User {
 /**
  * `GET /api/user/respawn-prohibited-rooms` response
  * @see {@link ScreepsHttpClient.userRespawnProhibitedRooms}
+ * @category HTTP API - User
  */
 export interface UserRespawnProhibitedRoomsResponse extends ScreepsResponse {
   /**
@@ -117,6 +147,7 @@ export interface UserRespawnProhibitedRoomsResponse extends ScreepsResponse {
 /**
  * `GET /api/user/rooms` response
  * @see {@link ScreepsHttpClient.userRooms}
+ * @category HTTP API - User
  */
 export interface UserRoomsResponse extends ScreepsResponse {
   /** All arrays in this object will always be empty */
@@ -128,6 +159,7 @@ export interface UserRoomsResponse extends ScreepsResponse {
 /**
  * `GET /api/user/stats` response
  * @see {@link ScreepsHttpClient.userStats}
+ * @category HTTP API - User
  */
 export interface UserStatsResponse extends ScreepsResponse {
   stats: {
@@ -139,6 +171,7 @@ export interface UserStatsResponse extends ScreepsResponse {
 /**
  * `GET /api/user/overview` response
  * @see {@link ScreepsHttpClient.userOverview}
+ * @category HTTP API - User
  */
 export interface UserOverviewResponse extends ScreepsResponse {
   statsMax: number
@@ -169,6 +202,7 @@ export interface UserOverviewResponse extends ScreepsResponse {
 /**
  * `GET /api/user/money-history` response
  * @see {@link ScreepsHttpClient.userMoneyHistory}
+ * @category HTTP API - User
  */
 export interface UserMoneyHistoryResponse extends ScreepsResponse {
   list: UserMoneyHistoryTransaction[]
@@ -177,7 +211,10 @@ export interface UserMoneyHistoryResponse extends ScreepsResponse {
   hasMore: boolean
 }
 
-/** A single record from {@link UserMoneyHistoryResponse} */
+/**
+ * A single record from {@link UserMoneyHistoryResponse}
+ * @category HTTP API - User
+ */
 export interface UserMoneyHistoryTransaction {
   _id: string
   /** ISO 8601 transaction timestamp */
@@ -200,6 +237,7 @@ export interface UserMoneyHistoryTransaction {
  * {@link UserMoneyHistoryTransaction.market} information about
  * an order update triggered via
  * {@link https://docs.screeps.com/api/#Game.market.changeOrderPrice | Game.market.changeOrderPrice()}.
+ * @category HTTP API - User
  */
 export interface MoneyHistoryChangeOrderPrice {
   changeOrderPrice: {
@@ -213,6 +251,7 @@ export interface MoneyHistoryChangeOrderPrice {
  * {@link UserMoneyHistoryTransaction.market} information about
  * an order update triggered via
  * {@link https://docs.screeps.com/api/#Game.market.extendOrder | Game.market.extendOrder()}.
+ * @category HTTP API - User
  */
 export interface MoneyHistoryExtendOrder {
   extendOrder: {
@@ -225,6 +264,7 @@ export interface MoneyHistoryExtendOrder {
  * {@link UserMoneyHistoryTransaction.market} details of
  * a market transaction executed via
  * {@link https://docs.screeps.com/api/#Game.market.deal | Game.market.deal()}.
+ * @category HTTP API - User
  */
 export interface MoneyHistoryFillOrder {
   resourceType: MarketResource
@@ -243,6 +283,7 @@ export interface MoneyHistoryFillOrder {
  * {@link UserMoneyHistoryTransaction.market} information about
  * a market {@link Order} created via
  * {@link https://docs.screeps.com/api/#Game.market.createOrder | Game.market.createOrder()}.
+ * @category HTTP API - User
  */
 export interface MoneyHistoryNewOrder {
   order: {
@@ -257,6 +298,7 @@ export interface MoneyHistoryNewOrder {
 /**
  * `POST /api/user/console` response
  * @see {@link ScreepsHttpClient.userConsole}
+ * @category HTTP API - User
  */
 export interface UserConsoleResponse extends ScreepsResponse {
   result: {
@@ -276,7 +318,53 @@ export interface UserConsoleResponse extends ScreepsResponse {
 /**
  * `GET /api/user/name` response
  * @see {@link ScreepsHttpClient.userName}
+ * @category HTTP API - User
  */
 export interface UserNameResponse extends ScreepsResponse {
   username: string
+}
+
+/**
+ * `GET /api/user/memory` response
+ * @see {@link ScreepsHttpClient.userMemoryGet}
+ * @category HTTP API - User
+ */
+export interface UserMemoryGetResponse extends ScreepsResponse {
+  /** Undefined if the specified memory path does not exist */
+  data?: unknown
+}
+
+/**
+ * `POST /api/user/memory` response
+ * @see {@link ScreepsHttpClient.userMemorySet}
+ * @category HTTP API - User
+ */
+export interface UserMemorySetResponse extends ScreepsResponse, ScreepsDbUpdateResponse {
+  ops: {
+    user: string
+    expression: string
+    hidden: boolean
+  }[]
+  data: string
+  insertedCount: number
+  insertedIds: string[]
+}
+
+/**
+ * `GET /api/user/memory-segment` response
+ * @see {@link ScreepsHttpClient.userMemorySegmentGet}
+ * @category HTTP API - User
+ */
+export interface UserMemorySegmentGetResponse extends ScreepsResponse {
+  /**
+   * The contents of the requested {@link https://docs.screeps.com/api/#RawMemory.segments | RawMemory.segments}.
+   *
+   * If a single segment ID is specified, returns the contents of that segment as a string.
+   *
+   * If multiple segment IDs are specified, returns the contents of each requested segment as a string array.
+   * The order of segment data in this array matches the order of the segment IDs array from the request.
+   *
+   * `null` will be returned instead of a string for any uninitialized segment.
+   */
+  data: string | null | (string | null)[]
 }
