+++ src/http/userDecorations.ts
@@ -2,13 +2,9 @@ import * as Decorations from '../common/decorations'
 import { ScreepsResponse } from './base'
 
 /**
- * Used with HTTP API endpoints on the `/api/user/decorations` path
- * @module
- */
-
-/**
  * `GET /api/user/decorations/inventory` response
  * @see {@link ScreepsHttpClient.userDecorationsInventory}
+ * @category HTTP API - User/Decorations
  */
 export interface UserDecorationInventoryResponse extends ScreepsResponse {
   list: Decorations.DecorationInstance[]
@@ -17,6 +13,7 @@ export interface UserDecorationInventoryResponse extends ScreepsResponse {
 /**
  * `GET /api/user/decorations/themes` response
  * @see {@link ScreepsHttpClient.userDecorationsThemes}
+ * @category HTTP API - User/Decorations
  */
 export interface UserDecorationThemesResponse extends ScreepsResponse {
   list: {
