+++ src/http/userMessages.ts
@@ -2,13 +2,9 @@ import { Users } from '../common/users'
 import { ScreepsDbUpdateResult, ScreepsResponse } from './base'
 
 /**
- * Used with HTTP API endpoints on the `/api/user/messages` path
- * @module
- */
-
-/**
  * `GET /api/user/messages/list` response
  * @see {@link ScreepsHttpClient.userMessagesList}
+ * @category HTTP API - User/Messages
  */
 export interface UserMessagesListResponse extends ScreepsResponse {
   messages: {
@@ -27,6 +23,7 @@ export interface UserMessagesListResponse extends ScreepsResponse {
 /**
  * A message result from {@link UserMessagesListResponse}
  * or {@link UserMessagesIndexResponse}
+ * @category HTTP API - User/Messages
  */
 export interface ListMessage {
   /** ID of the message */
@@ -43,6 +40,7 @@ export interface ListMessage {
 /**
  * `GET /api/user/messages/index` response
  * @see {@link ScreepsHttpClient.userMessagesIndex}
+ * @category HTTP API - User/Messages
  */
 export interface UserMessagesIndexResponse extends ScreepsResponse {
   messages: {
@@ -52,7 +50,10 @@ export interface UserMessagesIndexResponse extends ScreepsResponse {
   users: Users
 }
 
-/** A message result from {@link UserMessagesIndexResponse} */
+/**
+ * A message result from {@link UserMessagesIndexResponse}
+ * @category HTTP API - User/Messages
+ */
 export interface Message extends ListMessage {
   /** ID of the other user */
   respondent: string
@@ -65,6 +66,7 @@ export interface Message extends ListMessage {
 /**
  * `GET /api/user/messages/unread-count` response
  * @see {@link ScreepsHttpClient.userMessagesUnreadCount}
+ * @category HTTP API - User/Messages
  */
 export interface UserMessagesUnreadCountResponse extends ScreepsResponse {
   count: number
@@ -73,6 +75,7 @@ export interface UserMessagesUnreadCountResponse extends ScreepsResponse {
 /**
  * `POST /api/user/messages/mark-read` response
  * @see {@link ScreepsHttpClient.userMessagesMarkRead}
+ * @category HTTP API - User/Messages
  */
 export interface UserMessagesMarkReadResponse extends ScreepsResponse {
   0: number
