+++ src/socket/index.ts
@@ -1,9 +1,3 @@
 export * from './base'
 export * from './server'
 export * from './user'
-
-/**
- * Defines types/constants/enums/etc for the WebSocket API
- * @see {@link ScreepsSocketClient}
- * @module
- */
