+++ src/socket/user.ts
@@ -3,15 +3,11 @@ import { UserCodeModules } from '../common/users'
 import { SocketEvent } from './base'
 
 /**
- * Defines types/constants/enums/etc for `user` {@link SocketEvent}s
- * @module
- */
-
-/**
  * WebSocket event for code changes.
  *
  * When subscribed, the server sends an event with the full updated code base
  * every time the code is changed.
+ * @category WebSocket API - User
  */
 export interface UserCodeEvent extends SocketEvent {
   type: 'user'
@@ -19,7 +15,10 @@ export interface UserCodeEvent extends SocketEvent {
   data: UserCodeEventData
 }
 
-/** Payload of a {@link UserCodeEvent} */
+/**
+ * Payload of a {@link UserCodeEvent}
+ * @category WebSocket API - User
+ */
 export interface UserCodeEventData extends SocketEvent {
   /** Name of the updated branch */
   branch: string
@@ -36,6 +35,7 @@ export interface UserCodeEventData extends SocketEvent {
  *
  * When subscribed, the server sends one event per tick with console logs,
  * return values of commands, etc.
+ * @category WebSocket API - User
  */
 export interface UserConsoleEvent extends SocketEvent {
   type: 'user'
@@ -43,7 +43,10 @@ export interface UserConsoleEvent extends SocketEvent {
   data: UserConsoleEventData
 }
 
-/** Payload of a {@link UserConsoleEvent} */
+/**
+ * Payload of a {@link UserConsoleEvent}
+ * @category WebSocket API - User
+ */
 export interface UserConsoleEventData {
   /**
    * Console messages/results from the previous tick.
@@ -55,7 +58,10 @@ export interface UserConsoleEventData {
   shard?: string
 }
 
-/** Part of {@link UserConsoleEventData} */
+/**
+ * Part of {@link UserConsoleEventData}
+ * @category WebSocket API - User
+ */
 export interface UserConsoleMessages {
   /** Messages logged via `console.log()` */
   log: string[]
@@ -70,6 +76,7 @@ export interface UserConsoleMessages {
  * WebSocket event for CPU/memory usage updates.
  *
  * When subscribed, the server sends one event per tick (per shard?)
+ * @category WebSocket API - User
  */
 export interface UserCpuEvent extends SocketEvent {
   type: 'user'
@@ -77,7 +84,10 @@ export interface UserCpuEvent extends SocketEvent {
   data: UserCpuEventData
 }
 
-/** Payload of a {@link UserCpuEvent} */
+/**
+ * Payload of a {@link UserCpuEvent}
+ * @category WebSocket API - User
+ */
 export interface UserCpuEventData {
   /** CPU used last tick (integer value) */
   cpu: number
@@ -92,6 +102,7 @@ export interface UserCpuEventData {
  * is updated.
  * @example Raw sample events:
  * ["user:670e0c607317e200125d3aa2/resources",{"credits":0}]
+ * @category WebSocket API - User
  */
 export interface UserResourceEvent {
   type: 'user'
@@ -99,16 +110,25 @@ export interface UserResourceEvent {
   data: UserResourceEventData
 }
 
+/**
+ * Payload of a {@link UserResourceEvent}
+ * @category WebSocket API - User
+ */
 export type UserResourceEventData = {
   [resType in ResourceEventConstant]: number | undefined
 }
 
+/**
+ * A resource type that can be included in {@link UserResourceEventData}
+ * @category WebSocket API - User
+ */
 export type ResourceEventConstant = IntershardResource | 'credits'
 
 /**
  * WebSocket event that appears to be related to the Memory inspector watch list.
  * @example Raw sample events:
  * ["user:670e0c607317e200125d3aa2/memory/shardSeason/creeps.Scout-80965","undefined"]
+ * @category WebSocket API - User
  */
 export interface MemoryEvent {
   type: 'user'
